+++ python/lsst/rubintv/production/__init__.py
@@ -26,4 +26,5 @@ from .utils import *
 from .mountTorques import *
 from .monitorPlotting import *
 from .catchupService import *
-from .heartbeat import *
+from .uploadTools import *
+from .channels import *
