+++ python/lsst/rubintv/production/utils.py
@@ -27,15 +27,15 @@ from datetime import datetime, timedelta
 
 from lsst.summit.utils.butlerUtils import getSeqNumsForDayObs, makeDefaultLatissButler
 from lsst.summit.utils.utils import dayObsIntToString, setupLogging
-from . import CHANNELS, PREFIXES
-from . import (ImExaminerChannel,
-               SpecExaminerChannel,
-               MountTorqueChannel,
-               MonitorChannel,
-               MetadataServer,
-               Uploader,
-               )
-from .rubinTv import _dataIdToFilename
+from .channels import CHANNELS, PREFIXES
+from .rubinTv import (ImExaminerChannel,
+                      SpecExaminerChannel,
+                      MountTorqueChannel,
+                      MonitorChannel,
+                      MetadataServer,
+                      Uploader,
+                      _dataIdToFilename,
+                      )
 
 __all__ = ["checkRubinTvExternalPackages",
            "getPlotSeqNumsForDayObs",
